iot-core/requirements.txt
djangorestframework==3.9.2
markdown==2.6.11
Django==2.2
pytz==2018.5
django-filter==2.0.0
beautifulsoup4==4.6.3
requests==2.22.0
certifi==2018.8.24
psycopg2-binary==2.7.5
djangorestframework-simplejwt==3.2.3
httpsig==1.3.0
django-oauth-toolkit==1.2.0
django-cors-middleware==1.3.1
gunicorn==19.9.0